+++ extensions/copyq-clipboard-manager/src/view-clipboard-entries.tsx
@@ -1,34 +1,52 @@
 import {
-  ActionPanel,
-  Detail,
-  List,
-  Action,
   getPreferenceValues,
-  Icon,
-  Clipboard,
   openExtensionPreferences,
+  environment,
+  List,
+  Icon,
+  ActionPanel,
+  Action,
+  Clipboard,
+  Detail,
 } from "@raycast/api";
-import { useEffect, useState } from "react";
-import { exec } from "child_process";
+import { execSync, ExecSyncOptionsWithStringEncoding } from "child_process";
 import { accessSync, constants } from "fs";
 
 interface ExtensionPreferences {
   copyq_path: string;
   default_tab: string;
-  default_num_items: number;
+  max_items: number;
+}
+
+interface ClipboardItem {
+  row: number;
+  mimetypes: string[];
+  text: string;
 }
 
 export default function Command() {
+  // Handles preferences of the extension
   const preferences = getPreferenceValues<ExtensionPreferences>();
-  const copyqPath = preferences.copyq_path;
-  const defaultTab = preferences.default_tab;
-  let defaultNumItems = preferences.default_num_items;
+  const copyqPath = getPreferenceValues<ExtensionPreferences>().copyq_path;
+
+  // Set the default tab to preferences value if not passed from launchCommand
+  const selectedTab = environment.launchContext?.selectedTab;
+  const defaultTab = selectedTab !== undefined ? selectedTab : preferences.default_tab;
+
+  const maxItems = preferences.max_items;
+
+  // Define a function to get the amount of items in a tab
+  function numItems(tab: string) {
+    const command = `${copyqPath} tab ${tab} count`;
+    const stdout = execSync(command);
+    const numTabs = parseInt(stdout.toString());
+    return numTabs;
+  }
 
   // Check if CopyQ is installed
   try {
     accessSync(copyqPath, constants.X_OK);
   } catch (err) {
-    // Return error message
     return (
       <List>
         <List.Item
@@ -45,94 +63,78 @@ export default function Command() {
     );
   }
 
-  // Define a function that gets the amount of items in the tab
-  function getNumItems(): Promise<number> {
-    return new Promise((resolve, reject) => {
-      const command = `"${copyqPath}" tab ${defaultTab} count`;
-      exec(command, (error, stdout) => {
-        if (error) {
-          // Typically occurs when CopyQ path is incorrect or not installed.
-          console.log(`error: ${error.message}`);
-          reject(error);
-        } else {
-          resolve(parseInt(stdout, 10));
-        }
-      });
-    });
-  }
-
-  // Utilize the getNumItems function to check the number of items in the tab
-  async function checkNumItems() {
-    const numItems = await getNumItems();
-    // Ensure that defaultNumItems is less than or equal to the number of items in the tab
-    if (defaultNumItems > numItems) {
-      defaultNumItems = numItems;
-    }
-  }
-  checkNumItems();
-
-  const [clipboardContents, setClipboardContents] = useState<string[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  async function storeClipboardContents(tab: string, numItems: number) {
-    setIsLoading(true);
-    const promises: Promise<string>[] = [];
-
-    for (let i = 0; i < numItems; i++) {
-      const command = `"${copyqPath}" tab ${tab} read ${i}`;
-      const promise = new Promise<string>((resolve, reject) => {
-        exec(command, (error, stdout) => {
-          if (error) {
-            //Typically occurs when CopyQ path is incorrect or not installed.
-            console.log(`error: ${error.message}`);
-            reject(error);
-          } else {
-            resolve(stdout);
-          }
-        });
-      });
-
-      promises.push(promise);
+  // Define a function to get clipboard contents and return an array of [row, text]
+  function getClipboardContents(tab: string) {
+    // Determine correct amount of items to show
+    let amntItems = String(maxItems); // Displays maxItems items in the tab
+    if (maxItems > numItems(tab)) {
+      amntItems = String(numItems(tab)); // Does not display more items than there are in the tab
+    } else if (maxItems === 0) {
+      amntItems = "size()"; // Displays all items in the tab
     }
 
-    const results = await Promise.all(promises);
-    setClipboardContents(results);
-    setIsLoading(false);
+    // Define an ECMAScript to get clipboard contents
+    const script: string = `
+    tab('${tab}');
+    var result=[];
+    for ( var i = 0; i < ${amntItems}; ++i ) {
+        var obj = {} as Item;
+        obj.row = i;
+        obj.mimetypes = str(read("?", i)).split("\\n");
+        obj.mimetypes.pop();
+        obj.text = str(read(i));
+        result.push(obj);
+    }
+    JSON.stringify(result);
+    `;
+
+    const command: string = "copyq eval -";
+    const options: ExecSyncOptionsWithStringEncoding = {
+      input: script,
+      encoding: "utf8",
+    };
+
+    const stdout = execSync(command, options);
+    const jsonString = stdout.match(/\[.*\]/s); // Extract the JSON array part
+    const jsonArr = JSON.parse(jsonString![0]); // Parse the JSON array
+
+    // Return an array of [row, text]
+    const outputArray: Array<[number, string]> = jsonArr.map((item: ClipboardItem) => [item.row, item.text]);
+
+    return outputArray;
   }
 
+  // Define a function to select clipboard contents by row
   function selectClipboardContents(tab: string, index: number) {
     const command = `${copyqPath} tab ${tab} select ${index}`;
-    exec(command);
+    execSync(command);
   }
 
-  useEffect(() => {
-    async function fetchData() {
-      await checkNumItems();
-      storeClipboardContents(defaultTab, defaultNumItems);
-    }
+  // Get clipboard contents
+  const clipboardContents = getClipboardContents(defaultTab);
 
-    fetchData();
-  }, [defaultTab, defaultNumItems]);
+  return (
+    <List>
+      {clipboardContents.map(([row, text], index) => (
+        <List.Item
+          key={index}
+          title={text}
+          actions={
+            <ActionPanel>
+              <Action
+                title="Paste"
+                icon={Icon.Clipboard}
+                onAction={async () => {
+                  await selectClipboardContents(defaultTab, row);
+                  Clipboard.paste({ text });
+                }}
+              />
 
-  const items = clipboardContents.map((text, index) => (
-    <List.Item
-      key={index}
-      title={text}
-      actions={
-        <ActionPanel>
-          <Action
-            title="Paste"
-            icon={Icon.Clipboard}
-            onAction={async () => {
-              await selectClipboardContents(defaultTab, index);
-              Clipboard.paste({ text });
-            }}
-          />
-          <Action.Push title="Preview" icon={Icon.ArrowsExpand} target={<Detail markdown={text} />} />
-        </ActionPanel>
-      }
-    />
-  ));
-
-  return <List isLoading={isLoading}>{items}</List>;
+              <Action.Push title="Preview" icon={Icon.ArrowsExpand} target={<Detail markdown={text} />} />
+            </ActionPanel>
+          }
+        />
+      ))}
+    </List>
+  );
 }