+++ extensions/copyq-clipboard-manager/CHANGELOG.md
@@ -1,3 +1,10 @@
 # CopyQ Changelog
 
-## [Initial Version] - 2023-11-01
+## [Version 2.0] - 2023-11-18
+
+- Add second command to choose between a list of CopyQ tabs
+- Performance upgrades, allowing the user to display more clipboard contents faster.
+
+## [Added CopyQ Clipboard Manager] - 2023-11-01
+
+Initial version code